-cache & database logic encapuslated in dataManager object on all controllers

+++ Controllers/ActionsController.cs
@@ -26,16 +26,14 @@ namespace Services.Server.Controllers
     {
         private readonly IServicesFactory _factory;
         private readonly IScheduler _scheduler;
-        private readonly IDataAccess _dataAccess;
-        private readonly IDistributedCache _distributedCache;
+        private readonly IDataAccessManager _dataManager;
 
-        public ActionsController(IServicesFactory _factory, IScheduler _scheduler, IDataAccess _dataAccess, IDistributedCache distributedCache)
+        public ActionsController(IServicesFactory _factory, IScheduler _scheduler, IDataAccessManager dataAccess)
         {
 
             this._factory = _factory;
             this._scheduler = _scheduler;
-            this._dataAccess = _dataAccess;
-            _distributedCache = distributedCache;
+            _dataManager = dataAccess;
         }
 
         [Route("login")]
@@ -46,18 +44,11 @@ namespace Services.Server.Controllers
             {
                 data.Id = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
 
-                var UserServices = await _distributedCache.GetRecordAsync<List<UserServiceCredentials>>("userServices");
+                var UserServices = await _dataManager.GetAllUserServicesById(data);
 
                 if (UserServices is null)
                 {
-                    UserServices = await _dataAccess.GetAllUserServicesById(data.Id);
-
-                    if (UserServices is null)
-                    {
-                        return null;
-                    }
-                    //cache after retrieving from db 
-                    await _distributedCache.SetRecordAsync("userServices", UserServices);
+                    return null;
                 }
 
                 var singleService = UserServices.Where(a => a.Service == data.Service).FirstOrDefault();
@@ -85,34 +76,34 @@ namespace Services.Server.Controllers
         [HttpPost]
         public async Task<IDictionary<string, string>> GetImages(Data data)
         {
-            IDictionary<string, string> result = new Dictionary<string, string>();
+            IDictionary<string, string> images = new Dictionary<string, string>();
             try
             {
                 data.Id = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
                 IService service = _factory.GetService(data.Service);
 
-                var cachedImages = await _distributedCache.GetRecordAsync<IDictionary<string, string>>($"{data.Id}|images");
+                var cachedImages = await _dataManager.GetUserImages(data);
                 if (cachedImages is not null)
                 {
                     return cachedImages;
                 }
 
-                var userService = await _dataAccess.GetUserServiceByServiceNameAndId(data);
+                var userService = await _dataManager.GetUserServiceByServiceNameAndId(data);
                 data.UserName = userService.Username;
                 data.Password = userService.Password;
                 await Login(data);
 
-                result = await service.GetImages(data);
-                if (result.Count > 0)
+                images = await service.GetImages(data);
+                if (images.Count > 0)
                 {
-                    await _distributedCache.SetRecordAsync($"{data.Id}|images", result);
-                    return result;
+                    await _dataManager.SetUserImages(data, images);
+                    return images;
                 }
-                return result;
+                return images;
             }
             catch (Exception)
             {
-                return result;
+                return images;
             }
         }
 
@@ -120,12 +111,34 @@ namespace Services.Server.Controllers
         [HttpPost]
         public async Task<IDictionary<string, string>> RemoveImage(Data data)
         {
+
             IDictionary<string, string> result = new Dictionary<string, string>();
             try
             {
-                IService service = _factory.GetService(data.Service);
-                await Login(data);
+                data.Id = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+                IService service = default;
+
+                var userService = await _dataManager.GetUserServiceByServiceNameAndId(data);
+
+                service = _factory.GetService(data.Service);
+                if (userService is null)
+                {
+                    await service.AppStartUp(data);
+                    await Login(data);
+                }
+
+                var session = await _dataManager.GetServiceSession(new Data() { UserServiceId = userService.UserServiceId });
+                if (session is null)
+                {
+                    return result;
+                }
+                data.UserServiceId = session.SerivceId;
+                data.SessionId = session.SessionId;
+                data.HiddenUrl = session.HiddenUrl;
+
+
                 result = await service.RemoveImage(data);
+                await _dataManager.RemoveUserImage(data, result);
             }
             catch (Exception)
             {
@@ -184,7 +197,8 @@ namespace Services.Server.Controllers
                         Price = plan.Price,
                         Service = service.Service,
                         UserId = id,
-                        MessageId = Guid.NewGuid()
+                        MessageId = Guid.NewGuid(),
+                        Repeat = service.Repeat
                     });
                 }
             }
@@ -193,6 +207,28 @@ namespace Services.Server.Controllers
                 //log
             }
         }
+        //[Route("test")]
+        //[HttpGet]
+        //public async Task test()
+        //{
+        //    var id = "612e9a9bdc777780ab55c97a";
+        //    var UserServices = await _dataManager.GetRecordAsync<List<UserServiceCredentials>>($"userServices|{id}");
+        //    var data = new Data() { Id = id };
+        //    if (UserServices is null)
+        //    {
+        //        UserServices = await _dataManager.GetAllUserServicesById(data);
+
+        //        if (UserServices is null)
+        //        {
+
+        //        }
+        //        else
+        //        {
+        //            await _dataManager.SetRecordAsync($"userServices|{id}", UserServices);
+        //        }
+        //        //cache after retrieving from db 
+        //    }
+        //}
 
         [Route("authServices")]
         [HttpGet]
@@ -206,18 +242,11 @@ namespace Services.Server.Controllers
                 var data = new Data() { Id = id };
 
                 //check cache for services
-                var UserServices = await _distributedCache.GetRecordAsync<List<UserServiceCredentials>>("userServices");
+                var UserServices = await _dataManager.GetAllUserServicesById(data);
 
                 if (UserServices is null)
                 {
-                    UserServices = await _dataAccess.GetAllUserServicesById(id);
-
-                    if (UserServices is null)
-                    {
-                        return null;
-                    }
-                    //cache after retrieving from db 
-                    await _distributedCache.SetRecordAsync("userServices", UserServices);
+                    return null;
                 }
 
                 foreach (var singleService in UserServices)
@@ -231,37 +260,25 @@ namespace Services.Server.Controllers
 
                     await GetServicesSessions(servicesList, data, singleService, service);
                 }
-
             }
             catch (Exception)
             {
                 servicesList.Clear();
                 return servicesList;
-
             }
             return servicesList;
         }
 
         private async Task GetServicesSessions(List<Service> servicesList, Data data, UserServiceCredentials singleService, IService service)
         {
-            var chachedSession = await _distributedCache.GetRecordAsync<ServiceSessions>("serviceSession");
+            var chachedSession = await _dataManager.GetServiceSession(data);
 
             if (chachedSession is null)
             {
-                var session = await _dataAccess.CheckForServiceSession(data);
-
-                if (session is not null)
-                {
-                    servicesList.Add(singleService.Service);
-                    await _distributedCache.SetRecordAsync("serviceSession", session);
-                }
-                else
-                {
-                    // No session, log in to user and update session.
-                    data = await service.AppStartUp(data);
-                    await TryGetAndUpdateSession(data);
-                    servicesList.Add(singleService.Service);
-                }
+                // No session, log in to user and update session.
+                data = await service.AppStartUp(data);
+                await TryGetAndUpdateSession(data);
+                servicesList.Add(singleService.Service);
             }
             else
             {
@@ -270,25 +287,13 @@ namespace Services.Server.Controllers
         }
         private async Task GetSingleServiceSession(Data data, UserServiceCredentials singleService, ServiceSessions singleSession, IService service)
         {
-            var chachedSession = await _distributedCache.GetRecordAsync<ServiceSessions>("serviceSession");
+            var chachedSession = await _dataManager.GetServiceSession(data);
 
             if (chachedSession is null)
             {
-                var session = await _dataAccess.CheckForServiceSession(data);
-
-                if (session is not null)
-                {
-                    data.SessionId = session.SessionId;
-                    data.HiddenUrl = session.HiddenUrl;
-
-                    await _distributedCache.SetRecordAsync("serviceSession", session);
-                }
-                else
-                {
-                    // No session, log in to user and update session.
-                    var result = await service.AppStartUp(data);
-                    await TryGetAndUpdateSession(result);
-                }
+                // No session, log in to user and update session.
+                var result = await service.AppStartUp(data);
+                await TryGetAndUpdateSession(result);
             }
             else
             {
@@ -300,12 +305,12 @@ namespace Services.Server.Controllers
         {
             if (data.Result == Result.Success)
             {
-                await _dataAccess.UpdateServiceSession(data);
+                await _dataManager.UpdateServiceSession(data);
             }
             else
             {
                 //if unable to log into service, remove service and session.
-                await _dataAccess.RemoveServiceFromUser(data);
+                await _dataManager.RemoveServiceFromUser(data);
             }
         }
 
@@ -325,6 +330,16 @@ namespace Services.Server.Controllers
                 //log
                 return "unable to update";
             }
+        }
+
+        [Route("tutorial")]
+        [HttpPost]
+        public async Task SeenTutorial(Data data)
+        {
+            var id = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            data.Id = id;
+
+            await _dataManager.UpdateUser(data);
         }
     }
 }
